// db/database.types.ts
export interface Database {
// ...
  public: {
    Tables: {
// ...
      portfolio_groups: {
        Row: {
          created_at: string;
          deleted_at: string | null;
          description: string | null;
          id: string;
          name: string;
          portfolio_id: string;
          updated_at: string;
        };
        Insert: {
          created_at?: string;
          deleted_at?: string | null;
          description?: string | null;
          id?: string;
          name: string;
          portfolio_id: string;
          updated_at?: string;
        };
        Update: {
          created_at?: string;
          deleted_at?: string | null;
          description?: string | null;
          id?: string;
          name?: string;
          portfolio_id?: string;
          updated_at?: string;
        };
        Relationships: [
          {
            foreignKeyName: "portfolio_groups_portfolio_id_fkey";
            columns: ["portfolio_id"];
            isOneToOne: false;
            referencedRelation: "portfolio_summary";
            referencedColumns: ["portfolio_id"];
          },
          {
            foreignKeyName: "portfolio_groups_portfolio_id_fkey";
            columns: ["portfolio_id"];
            isOneToOne: false;
            referencedRelation: "portfolios";
            referencedColumns: ["id"];
          }
        ];
      };
      portfolios: {
        Row: {
          id: string;
          user_id: string;
          name: string;
          description: string | null;
          created_at: string;
          updated_at: string;
          deleted_at: string | null;
        };
        Insert: {
          id?: string;
          user_id: string;
          name: string;
          description?: string | null;
          created_at?: string;
          updated_at?: string;
          deleted_at?: string | null;
        };
        Update: {
          id?: string;
          user_id?: string;
          name?: string;
          description?: string | null;
          created_at?: string;
          updated_at?: string;
          deleted_at?: string | null;
        };
        Relationships: [
          {
            foreignKeyName: "portfolios_user_id_fkey";
            columns: ["user_id"];
            referencedRelation: "users";
            referencedColumns: ["id"];
          }
        ];
      };
// ...
      positions: {
        Row: {
          average_buy_price: number;
          break_even_price: number | null;
          created_at: string;
          cryptocurrency_id: string;
          deleted_at: string | null;
          id: string;
          portfolio_id: string;
          total_amount: number;
          updated_at: string;
        };
        Insert: {
          average_buy_price: number;
          break_even_price?: number | null;
          created_at?: string;
          cryptocurrency_id: string;
          deleted_at?: string | null;
          id?: string;
          portfolio_id: string;
          total_amount: number;
          updated_at?: string;
        };
        Update: {
          average_buy_price?: number;
          break_even_price?: number | null;
          created_at?: string;
          cryptocurrency_id?: string;
          deleted_at?: string | null;
          id?: string;
          portfolio_id?: string;
          total_amount?: number;
          updated_at?: string;
        };
        Relationships: [
          {
            foreignKeyName: "positions_portfolio_id_fkey";
            columns: ["portfolio_id"];
            isOneToOne: false;
            referencedRelation: "portfolio_summary";
            referencedColumns: ["portfolio_id"];
          },
          {
            foreignKeyName: "positions_portfolio_id_fkey";
            columns: ["portfolio_id"];
            isOneToOne: false;
            referencedRelation: "portfolios";
            referencedColumns: ["id"];
          }
        ];
      };
// ...
    };
// ...
  };
// ...
}